extract server code to another file
